requirements.txt
streamlit>=1.33.0
pydantic>=2.7.0
tenacity>=8.3.0
python-dotenv>=1.0.1
requests>=2.32.0
